Refactor project structure: - Added .gitattributes for text file handling. - Updated .gitignore to exclude additional files and directories. - Removed unused routes and views related to users, schedules, hotels, and vegetables. - Updated package dependencies to version 6.2.1 for Prisma and related packages. - Modified app.js to streamline routing. - Updated README with setup instructions for the World Weather Application. - Removed obsolete database files and migrations.

+++ prisma/migrations/20250123002320_init/migration.sql
@@ -5,6 +5,6 @@ CREATE TABLE "WeatherReport" (
     "windDirection" TEXT NOT NULL,
     "windStrength" TEXT NOT NULL,
     "weather" TEXT NOT NULL,
-    "pressure" REAL NOT NULL,
+    "pressure" INTEGER NOT NULL,
     "temperature" REAL NOT NULL
 );
